rename messages/setup.py → messages/session_setup.py
GitHub's auto-detected Dependency Graph submission misidentifies
aiomoqt/messages/setup.py as a Python package install script and
runs a "Graph Update: pip in /., /aiomoqt/messages" workflow on
every push to main. Our file is the MoQT session-lifecycle message
module (ClientSetup, ServerSetup, GoAway) — not a setup.py in the
packaging sense.

Two file changes; zero public-API impact (callers do
`from aiomoqt.messages import ClientSetup`, never reference the
submodule directly).
